fix: added a TODO Comment for constants - added TODO Comment in title and Text
