+++ pokedex_async_redux/.gitignore
@@ -13,7 +13,8 @@
 .history
 .svn/
 migrate_working_dir/
-.idea/
+*.idea/
+**/*.idea
 
 # IntelliJ related
 *.iml
