+++ pokedex_async_redux/analysis_options.yaml
@@ -24,6 +24,12 @@ linter:
   rules:
     # avoid_print: false  # Uncomment to disable the `avoid_print` rule
     # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
+analyzer:
+  exclude:
+    -"**/*.g.dart"
+    -"**/*.freezed.dart"
 
+  errors:
+    invalid_annotation_target: ignore
 # Additional information about this file can be found at
 # https://dart.dev/guides/language/analysis-options
